Move MM_VERSION and MM_DATETIME to mm2 To avoid rebuilding common crate every time
