+++ Cargo.toml
@@ -131,6 +131,11 @@ rand4 = { version = "0.4", package = "rand" }
 testcontainers = { git = "https://github.com/artemii235/testcontainers-rs.git" }
 winapi = "0.3"
 
+[build-dependencies]
+chrono = "0.4"
+gstuff = { version = "0.6", features = ["nightly", "term"] }
+regex = "1"
+
 [workspace]
 members = [
     "mm2src/coins",
