+++ mm2src/common/Cargo.toml
@@ -76,19 +76,6 @@ rusqlite = { version = "0.24.2", features = ["bundled"] }
 tokio = { version = "0.2.22", features = ["io-util", "rt-threaded", "stream", "tcp"] }
 
 [build-dependencies]
-bzip2 = "0.3"
 cc = "1.0"
-chrono = "0.4"
 fomat-macros = "0.3"
-futures = { version = "0.3", package = "futures" }
-futures01 = { version = "0.1", package = "futures" }
-futures-cpupool = "0.1"
-glob = "0.3"
 gstuff = { version = "0.6", features = ["nightly", "term"] }
-itertools = "0.8"
-libflate = "0.1"
-num_cpus = "1.10"
-regex = "1"
-sysinfo = "0.9"
-tar = "0.4"
-winapi = "0.3"
