+++ mm2src/rpc/lp_commands.rs
@@ -22,7 +22,7 @@ use coins::{disable_coin as disable_coin_impl, lp_coinfind, lp_coininit, MmCoinE
 use common::executor::{spawn, Timer};
 use common::mm_ctx::MmArc;
 use common::mm_metrics::MetricsOps;
-use common::{rpc_err_response, rpc_response, HyRes, MM_DATETIME, MM_VERSION};
+use common::{rpc_err_response, rpc_response, HyRes};
 use futures::compat::Future01CompatExt;
 use http::Response;
 use serde_json::{self as json, Value as Json};
@@ -30,6 +30,7 @@ use std::borrow::Cow;
 
 use crate::mm2::lp_ordermatch::{cancel_orders_by, CancelBy};
 use crate::mm2::lp_swap::active_swaps_using_coin;
+use crate::mm2::{MM_DATETIME, MM_VERSION};
 
 /// Attempts to disable the coin
 pub async fn disable_coin(ctx: MmArc, req: Json) -> Result<Response<Vec<u8>>, String> {
